fix(CI): BufferSource casts para Web Crypto (TS 5.6/5.7 compat)
- master-identity.ts: fromB64url → BufferSource en s.sign/verify
- relay-crypto.ts: fromB64 → BufferSource en importKey/decrypt
- keyStorage.ts: fromB64 → Uint8Array (TS 5.6), encrypt/decrypt → BufferSource
- audio-analyzer.ts: Engine.data → Uint8Array, getByteFrequencyData(data) sin cast
- aurora-orb-bus.ts: MicShared freq/time → Uint8Array

Estrategia: usar  en llamadas Web Crypto (compat TS 5.6 + 5.7),
tipar / como  plano (TS 5.6 compatible).
Build verificado: exit 0, 105/105 páginas.

+++ src/lib/audiomorphic/audio-analyzer.ts
@@ -265,7 +265,7 @@ export function getMetrics(sensitivity: number, freqRange: number): AudioMetrics
     if (!analyser || !data || engine.state !== "live") return SILENT_METRICS;
 
     // El tipo de `getByteFrequencyData` varía entre libs de TS (Uint8Array).
-    analyser.getByteFrequencyData(data as Uint8Array);
+    analyser.getByteFrequencyData(data);
 
     const rangeLimit = Math.floor(data.length * freqRange);
     let totalMagnitude = 0;
